chore: remove /v1/runs endpoints and runs infrastructure
