--- assistant/src/__tests__/runtime-runs-http.test.ts
@@ -1,527 +1,0 @@
-/**
- * HTTP-layer integration tests for the run API endpoints.
- *
- * Tests POST /runs, GET /runs/:id, and POST /runs/:id/decision
- * through RuntimeHttpServer with a real RunOrchestrator instance.
- */
-import { mkdtempSync, realpathSync,rmSync } from 'node:fs';
-import { tmpdir } from 'node:os';
-import { join } from 'node:path';
-
-import { afterAll, beforeEach, describe, expect, mock,test } from 'bun:test';
-
-import type { ServerMessage } from '../daemon/ipc-protocol.js';
-import type { Session } from '../daemon/session.js';
-
-const testDir = realpathSync(mkdtempSync(join(tmpdir(), 'runtime-runs-http-test-')));
-
-mock.module('../util/platform.js', () => ({
-  getRootDir: () => testDir,
-  getDataDir: () => testDir,
-  isMacOS: () => process.platform === 'darwin',
-  isLinux: () => process.platform === 'linux',
-  isWindows: () => process.platform === 'win32',
-  getSocketPath: () => join(testDir, 'test.sock'),
-  getPidPath: () => join(testDir, 'test.pid'),
-  getDbPath: () => join(testDir, 'test.db'),
-  getLogPath: () => join(testDir, 'test.log'),
-  ensureDataDir: () => {},
-}));
-
-mock.module('../util/logger.js', () => ({
-  getLogger: () => new Proxy({} as Record<string, unknown>, {
-    get: () => () => {},
-  }),
-}));
-
-mock.module('../config/loader.js', () => ({
-  getConfig: () => ({
-    model: 'test',
-    provider: 'test',
-    apiKeys: {},
-    memory: { enabled: false },
-    rateLimit: { maxRequestsPerMinute: 0, maxTokensPerSession: 0 },
-    secretDetection: { enabled: false },
-  }),
-}));
-
-import { getDb, initializeDb, resetDb } from '../memory/db.js';
-import { RuntimeHttpServer } from '../runtime/http-server.js';
-import { RunOrchestrator } from '../runtime/run-orchestrator.js';
-
-initializeDb();
-
-// ---------------------------------------------------------------------------
-// Session helpers
-// ---------------------------------------------------------------------------
-
-function makeCompletingSession(): Session {
-  let processing = false;
-  return {
-    isProcessing: () => processing,
-    persistUserMessage: () => undefined as unknown as string,
-    memoryPolicy: { scopeId: 'default', includeDefaultFallback: false, strictSideEffects: false },
-    setChannelCapabilities: () => {},
-    setAssistantId: () => {},
-    setGuardianContext: () => {},
-    setCommandIntent: () => {},
-    setTurnChannelContext: () => {},
-    setTurnInterfaceContext: () => {},
-    setVoiceCallControlPrompt: () => {},
-    updateClient: () => {},
-    runAgentLoop: async () => {
-      processing = true;
-      await new Promise((r) => setTimeout(r, 20));
-      processing = false;
-    },
-    handleConfirmationResponse: () => {},
-    handleSecretResponse: () => {},
-  } as unknown as Session;
-}
-
-function makeFailingSession(errorMsg: string): Session {
-  return {
-    isProcessing: () => false,
-    persistUserMessage: () => undefined as unknown as string,
-    memoryPolicy: { scopeId: 'default', includeDefaultFallback: false, strictSideEffects: false },
-    setChannelCapabilities: () => {},
-    setAssistantId: () => {},
-    setGuardianContext: () => {},
-    setCommandIntent: () => {},
-    setTurnChannelContext: () => {},
-    setTurnInterfaceContext: () => {},
-    setVoiceCallControlPrompt: () => {},
-    updateClient: () => {},
-    runAgentLoop: async (_content: string, _messageId: string, onEvent: (msg: ServerMessage) => void) => {
-      onEvent({ type: 'error', message: errorMsg });
-    },
-    handleConfirmationResponse: () => {},
-    handleSecretResponse: () => {},
-  } as unknown as Session;
-}
-
-function makeConfirmationSession(toolName: string): Session {
-  let clientHandler: (msg: ServerMessage) => void = () => {};
-  return {
-    isProcessing: () => false,
-    persistUserMessage: () => undefined as unknown as string,
-    memoryPolicy: { scopeId: 'default', includeDefaultFallback: false, strictSideEffects: false },
-    setChannelCapabilities: () => {},
-    setAssistantId: () => {},
-    setGuardianContext: () => {},
-    setCommandIntent: () => {},
-    setTurnChannelContext: () => {},
-    setTurnInterfaceContext: () => {},
-    setVoiceCallControlPrompt: () => {},
-    updateClient: (handler: (msg: ServerMessage) => void) => {
-      clientHandler = handler;
-    },
-    runAgentLoop: async () => {
-      clientHandler({
-        type: 'confirmation_request',
-        requestId: 'req-1',
-        toolName,
-        input: { objective: 'test task' },
-        riskLevel: 'medium',
-        allowlistOptions: [],
-        scopeOptions: [],
-      });
-      // Hang to simulate waiting for decision
-      await new Promise<void>(() => {});
-    },
-    handleConfirmationResponse: () => {},
-    handleSecretResponse: () => {},
-  } as unknown as Session;
-}
-
-function makeHangingSession(): Session {
-  let processing = false;
-  return {
-    isProcessing: () => processing,
-    persistUserMessage: () => undefined as unknown as string,
-    memoryPolicy: { scopeId: 'default', includeDefaultFallback: false, strictSideEffects: false },
-    setChannelCapabilities: () => {},
-    setAssistantId: () => {},
-    setGuardianContext: () => {},
-    setCommandIntent: () => {},
-    setTurnChannelContext: () => {},
-    setTurnInterfaceContext: () => {},
-    setVoiceCallControlPrompt: () => {},
-    updateClient: () => {},
-    runAgentLoop: async () => {
-      processing = true;
-      await new Promise<void>(() => {});
-    },
-    handleConfirmationResponse: () => {},
-    handleSecretResponse: () => {},
-  } as unknown as Session;
-}
-
-// ---------------------------------------------------------------------------
-// Tests
-// ---------------------------------------------------------------------------
-
-const TEST_TOKEN = 'test-bearer-token-runs';
-const AUTH_HEADERS = { Authorization: `Bearer ${TEST_TOKEN}` };
-
-describe('runtime runs — HTTP layer', () => {
-  let server: RuntimeHttpServer;
-  let port: number;
-
-  beforeEach(() => {
-    const db = getDb();
-    db.run('DELETE FROM message_runs');
-    db.run('DELETE FROM messages');
-    db.run('DELETE FROM conversations');
-    db.run('DELETE FROM conversation_keys');
-  });
-
-  afterAll(() => {
-    resetDb();
-    try { rmSync(testDir, { recursive: true, force: true }); } catch { /* best effort */ }
-  });
-
-  async function startServer(sessionFactory: () => Session): Promise<{ orchestrator: RunOrchestrator }> {
-    port = 18000 + Math.floor(Math.random() * 1000);
-    const orchestrator = new RunOrchestrator({
-      getOrCreateSession: async () => sessionFactory(),
-      resolveAttachments: () => [],
-      deriveDefaultStrictSideEffects: () => false,
-    });
-    server = new RuntimeHttpServer({ port, bearerToken: TEST_TOKEN, runOrchestrator: orchestrator });
-    await server.start();
-    return { orchestrator };
-  }
-
-  async function stopServer(): Promise<void> {
-    await server?.stop();
-  }
-
-  function runsUrl(path = ''): string {
-    return `http://127.0.0.1:${port}/v1/runs${path}`;
-  }
-
-  // ── Auth ────────────────────────────────────────────────────────────
-
-  test('returns 401 when bearer token is missing', async () => {
-    await startServer(() => makeCompletingSession());
-
-    const res = await fetch(runsUrl(), {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({
-        conversationKey: 'conv-noauth',
-        content: 'Hi',
-        sourceChannel: 'vellum',
-        interface: 'macos',
-      }),
-    });
-
-    expect(res.status).toBe(401);
-    await stopServer();
-  });
-
-  test('returns 401 when bearer token is wrong', async () => {
-    await startServer(() => makeCompletingSession());
-
-    const res = await fetch(runsUrl(), {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json', Authorization: 'Bearer wrong-token' },
-      body: JSON.stringify({
-        conversationKey: 'conv-badauth',
-        content: 'Hi',
-        sourceChannel: 'vellum',
-        interface: 'macos',
-      }),
-    });
-
-    expect(res.status).toBe(401);
-    await stopServer();
-  });
-
-  test('healthz is accessible without auth and includes disk space info', async () => {
-    await startServer(() => makeCompletingSession());
-
-    const res = await fetch(`http://127.0.0.1:${port}/healthz`);
-    expect(res.status).toBe(200);
-
-    const body = await res.json() as { status: string; disk: { path: string; totalMb: number; usedMb: number; freeMb: number } | null };
-    expect(body.status).toBe('healthy');
-    if (body.disk != null) {
-      expect(typeof body.disk.path).toBe('string');
-      expect(body.disk.totalMb).toBeGreaterThan(0);
-      expect(body.disk.usedMb).toBeGreaterThanOrEqual(0);
-      expect(body.disk.freeMb).toBeGreaterThanOrEqual(0);
-    }
-
-    await stopServer();
-  });
-
-  // ── POST /runs ──────────────────────────────────────────────────────
-
-  test('POST /runs creates a run and returns 201', async () => {
-    await startServer(() => makeCompletingSession());
-
-    const res = await fetch(runsUrl(), {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json', ...AUTH_HEADERS },
-      body: JSON.stringify({
-        conversationKey: 'conv-1',
-        content: 'Hello',
-        sourceChannel: 'vellum',
-        interface: 'macos',
-      }),
-    });
-    const body = await res.json() as { id: string; status: string; messageId: string; createdAt: string };
-
-    expect(res.status).toBe(201);
-    expect(body.id).toBeDefined();
-    expect(body.status).toBe('running');
-    expect(body.messageId).toBeNull();
-    expect(body.createdAt).toBeDefined();
-
-    await stopServer();
-  });
-
-  test('POST /runs returns 400 when conversationKey missing', async () => {
-    await startServer(() => makeCompletingSession());
-
-    const res = await fetch(runsUrl(), {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json', ...AUTH_HEADERS },
-      body: JSON.stringify({
-        content: 'Hello',
-        sourceChannel: 'vellum',
-        interface: 'macos',
-      }),
-    });
-
-    expect(res.status).toBe(400);
-    const body = await res.json() as { error: string };
-    expect(body.error).toContain('conversationKey');
-
-    await stopServer();
-  });
-
-  test('POST /runs returns 400 when content is empty', async () => {
-    await startServer(() => makeCompletingSession());
-
-    const res = await fetch(runsUrl(), {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json', ...AUTH_HEADERS },
-      body: JSON.stringify({
-        conversationKey: 'conv-2',
-        content: '',
-        sourceChannel: 'vellum',
-        interface: 'macos',
-      }),
-    });
-
-    expect(res.status).toBe(400);
-
-    await stopServer();
-  });
-
-  test('POST /runs returns 409 when session busy', async () => {
-    const session = makeHangingSession();
-    await startServer(() => session);
-
-    // First run starts and hangs
-    const res1 = await fetch(runsUrl(), {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json', ...AUTH_HEADERS },
-      body: JSON.stringify({
-        conversationKey: 'conv-busy',
-        content: 'First',
-        sourceChannel: 'vellum',
-        interface: 'macos',
-      }),
-    });
-    expect(res1.status).toBe(201);
-
-    await new Promise((r) => setTimeout(r, 30));
-
-    // Second run should be rejected
-    const res2 = await fetch(runsUrl(), {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json', ...AUTH_HEADERS },
-      body: JSON.stringify({
-        conversationKey: 'conv-busy',
-        content: 'Second',
-        sourceChannel: 'vellum',
-        interface: 'macos',
-      }),
-    });
-    expect(res2.status).toBe(409);
-
-    await stopServer();
-  });
-
-  // ── GET /runs/:id ───────────────────────────────────────────────────
-
-  test('GET /runs/:id returns run status', async () => {
-    await startServer(() => makeHangingSession());
-
-    const createRes = await fetch(runsUrl(), {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json', ...AUTH_HEADERS },
-      body: JSON.stringify({
-        conversationKey: 'conv-get',
-        content: 'Test',
-        sourceChannel: 'vellum',
-        interface: 'macos',
-      }),
-    });
-    const { id } = await createRes.json() as { id: string };
-
-    const getRes = await fetch(runsUrl(`/${id}`), { headers: AUTH_HEADERS });
-    const body = await getRes.json() as { id: string; status: string; messageId: string };
-
-    expect(getRes.status).toBe(200);
-    expect(body.id).toBe(id);
-    expect(body.status).toBe('running');
-
-    await stopServer();
-  });
-
-  test('GET /runs/:id returns completed after agent loop finishes', async () => {
-    await startServer(() => makeCompletingSession());
-
-    const createRes = await fetch(runsUrl(), {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json', ...AUTH_HEADERS },
-      body: JSON.stringify({
-        conversationKey: 'conv-done',
-        content: 'Build it',
-        sourceChannel: 'vellum',
-        interface: 'macos',
-      }),
-    });
-    const { id } = await createRes.json() as { id: string };
-
-    await new Promise((r) => setTimeout(r, 100));
-
-    const getRes = await fetch(runsUrl(`/${id}`), { headers: AUTH_HEADERS });
-    const body = await getRes.json() as { id: string; status: string };
-
-    expect(getRes.status).toBe(200);
-    expect(body.status).toBe('completed');
-
-    await stopServer();
-  });
-
-  test('GET /runs/:id returns failed with error', async () => {
-    await startServer(() => makeFailingSession('Backend error'));
-
-    const createRes = await fetch(runsUrl(), {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json', ...AUTH_HEADERS },
-      body: JSON.stringify({
-        conversationKey: 'conv-fail',
-        content: 'Do it',
-        sourceChannel: 'vellum',
-        interface: 'macos',
-      }),
-    });
-    const { id } = await createRes.json() as { id: string };
-
-    await new Promise((r) => setTimeout(r, 50));
-
-    const getRes = await fetch(runsUrl(`/${id}`), { headers: AUTH_HEADERS });
-    const body = await getRes.json() as { id: string; status: string; error: string };
-
-    expect(getRes.status).toBe(200);
-    expect(body.status).toBe('failed');
-    expect(body.error).toBe('Backend error');
-
-    await stopServer();
-  });
-
-  test('GET /runs/:id returns 404 for unknown run', async () => {
-    await startServer(() => makeCompletingSession());
-
-    const res = await fetch(runsUrl('/nonexistent'), { headers: AUTH_HEADERS });
-    expect(res.status).toBe(404);
-
-    await stopServer();
-  });
-
-  // ── POST /runs/:id/decision ─────────────────────────────────────────
-
-  test('POST /runs/:id/decision returns accepted for pending confirmation', async () => {
-    await startServer(() => makeConfirmationSession('swarm_delegate'));
-
-    const createRes = await fetch(runsUrl(), {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json', ...AUTH_HEADERS },
-      body: JSON.stringify({
-        conversationKey: 'conv-decide',
-        content: 'Approve',
-        sourceChannel: 'vellum',
-        interface: 'macos',
-      }),
-    });
-    const { id } = await createRes.json() as { id: string };
-
-    await new Promise((r) => setTimeout(r, 50));
-
-    // Verify pending state via GET
-    const getRes = await fetch(runsUrl(`/${id}`), { headers: AUTH_HEADERS });
-    const runBody = await getRes.json() as { status: string; pendingConfirmation: { toolName: string } };
-    expect(runBody.status).toBe('needs_confirmation');
-    expect(runBody.pendingConfirmation.toolName).toBe('swarm_delegate');
-
-    // Submit decision
-    const decisionRes = await fetch(runsUrl(`/${id}/decision`), {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json', ...AUTH_HEADERS },
-      body: JSON.stringify({ decision: 'allow' }),
-    });
-    const decisionBody = await decisionRes.json() as { accepted: boolean };
-
-    expect(decisionRes.status).toBe(200);
-    expect(decisionBody.accepted).toBe(true);
-
-    await stopServer();
-  });
-
-  test('POST /runs/:id/decision returns 400 for invalid decision', async () => {
-    await startServer(() => makeHangingSession());
-
-    const createRes = await fetch(runsUrl(), {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json', ...AUTH_HEADERS },
-      body: JSON.stringify({
-        conversationKey: 'conv-bad-dec',
-        content: 'Test',
-        sourceChannel: 'vellum',
-        interface: 'macos',
-      }),
-    });
-    const { id } = await createRes.json() as { id: string };
-
-    const res = await fetch(runsUrl(`/${id}/decision`), {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json', ...AUTH_HEADERS },
-      body: JSON.stringify({ decision: 'maybe' }),
-    });
-
-    expect(res.status).toBe(400);
-
-    await stopServer();
-  });
-
-  test('POST /runs/:id/decision returns 404 for unknown run', async () => {
-    await startServer(() => makeCompletingSession());
-
-    const res = await fetch(runsUrl('/nonexistent/decision'), {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json', ...AUTH_HEADERS },
-      body: JSON.stringify({ decision: 'allow' }),
-    });
-
-    expect(res.status).toBe(404);
-
-    await stopServer();
-  });
-});
