+++ assistant/src/memory/schema.ts
@@ -224,26 +224,6 @@ export const channelInboundEvents = sqliteTable('channel_inbound_events', {
   updatedAt: integer('updated_at').notNull(),
 });
 
-// ── Message Runs (approval flow) ─────────────────────────────────────
-
-export const messageRuns = sqliteTable('message_runs', {
-  id: text('id').primaryKey(),
-  conversationId: text('conversation_id')
-    .notNull()
-    .references(() => conversations.id, { onDelete: 'cascade' }),
-  messageId: text('message_id')
-    .references(() => messages.id, { onDelete: 'cascade' }),
-  status: text('status').notNull().default('running'),          // running | needs_confirmation | needs_secret | completed | failed
-  pendingConfirmation: text('pending_confirmation'),            // JSON when status=needs_confirmation
-  pendingSecret: text('pending_secret'),                        // JSON when status=needs_secret
-  inputTokens: integer('input_tokens').notNull().default(0),
-  outputTokens: integer('output_tokens').notNull().default(0),
-  estimatedCost: real('estimated_cost').notNull().default(0),
-  error: text('error'),
-  createdAt: integer('created_at').notNull(),
-  updatedAt: integer('updated_at').notNull(),
-});
-
 export const memoryCheckpoints = sqliteTable('memory_checkpoints', {
   key: text('key').primaryKey(),
   value: text('value').notNull(),
